reorganize helper functions out of common.ts
